requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
python-dotenv==1.0.1
flask==3.0.2
gunicorn==21.2.0
requests==2.31.0